[processing]changed native raster format for SAGA algorithms from tif to sdat
This eliminates the need to export to a QGIS compatible format after generating a raster output. The SDAT format is the native SAGA format, and it is supported by GDAL, so it is also supported by QGIS. This was not used before, but now we use that to reduce the nuber of steps and provide a better integration.

Conversion to other formats is now done by Processing itself, not by the SAGA algorithm.

fixes #10735